+++ src/Console/Input/Input.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+namespace Bartlett\PHPToolbox\Console\Input;
+
+use Symfony\Component\Console\Input\ArgvInput;
+
+/**
+ * @since Release 1.0.0alpha2
+ */
+final class Input extends ArgvInput
+{
+}