+++ src/Console/Output/Output.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+namespace Bartlett\PHPToolbox\Console\Output;
+
+use Symfony\Component\Console\Output\ConsoleOutput;
+
+/**
+ * @since Release 1.0.0alpha2
+ */
+final class Output extends ConsoleOutput
+{
+}