+++ src/Event/EventDispatcher.php
@@ -1,0 +1,24 @@
+<?php declare(strict_types=1);
+
+namespace Bartlett\PHPToolbox\Event;
+
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\EventDispatcher\EventDispatcher as SymfonyEventDispatcher;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+
+/**
+ * @since Release 1.0.0alpha2
+ */
+final class EventDispatcher extends SymfonyEventDispatcher
+{
+    public function __construct(
+        InputInterface $input,
+        EventSubscriberInterface $profileEventSubscriber
+    ) {
+        parent::__construct();
+
+        if ($input->hasParameterOption('--profile')) {
+            $this->addSubscriber($profileEventSubscriber);
+        }
+    }
+}